# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Tài liệu API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/users/search": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Tìm kiếm người dùng theo từ khóa",
                "description": "API này cho phép tìm kiếm người dùng theo từ khóa trong tên hoặc email.",
                "operationId": "searchUsers",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm người dùng",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "John"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách người dùng tìm thấy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "../components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}